+++ Dionysos/Services/VendorServices/VendorDeletionService.cs
@@ -14,10 +14,7 @@ public class VendorDeletionService
     public void DeleteVendor(int id)
     {
         var vendor = _dbContext.Vendors.SingleOrDefault(x => x.Id == id);
-        if (vendor is not null)
-        {
-            _dbContext.Vendors.Remove(vendor);
-        }
+        if (vendor is not null) _dbContext.Vendors.Remove(vendor);
         _dbContext.SaveChanges();
     }
 }
